\name{vfranking-package}
\alias{vfranking-package}
\alias{vfranking}
\docType{package}
\title{
Value function based ranking
}
\description{
Implementation of multiple criteria ranking methods that use additive value function to model preferences.
}
\details{
\tabular{ll}{
Package: \tab vfranking\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-10-21\cr
}
}
\author{
Krzysztof Ciomek

Maintainer: Krzysztof Ciomek <k.ciomek at gmail.com>
}
\keyword{ additive mdca ordinal regression robust ror ranking uta value vfranking }